Split unit and integration tests
