+++ test/integration/meson.build
@@ -1,9 +1,26 @@
 if get_option('enable_tests')
+  test_conf_data = configuration_data()
+  test_conf_data.set_quoted('RIZIN_BUILD_PATH', rizin_exe.full_path())
+  test_config_h = configure_file(
+    input: 'test_config.h.in',
+    output: 'test_config.h',
+    configuration: test_conf_data,
+  )
+
   tests = [
+    'analysis_il',
+    'autocmplt',
+    'bin',
     'bin_vfiles',
     'cpu_platform_profiles',
+    'dwarf',
+    'dwarf_info',
+    'dwarf_integration',
     'open_analyse_save_load_project',
-    'analysis_il',
+    'pdb',
+    'project_migrate',
+    'rzpipe',
+    'sign_integration',
   ]
 
   unit_test_env = environment()
@@ -14,7 +31,8 @@ if get_option('enable_tests')
   foreach test : tests
     exe = executable('test_@0@'.format(test), 'test_@0@.c'.format(test),
       c_args: executable_cflags,
-      include_directories: [platform_inc],
+      # '.' to find autogenerated "test_config.h"
+      include_directories: [platform_inc, '.'],
       dependencies: [
         rz_util_dep,
         rz_main_dep,
@@ -45,6 +63,6 @@ if get_option('enable_tests')
       implicit_include_directories: false,
       link_args: executable_linkflags
     )
-    test(test, exe, workdir: join_paths(meson.current_source_dir(), '..'), env: unit_test_env)
+    test(test, exe, workdir: join_paths(meson.current_source_dir(), '..'), env: unit_test_env, suite: 'integration')
   endforeach
 endif
