+++ test/unit/auxiliary/meson.build
@@ -6,10 +6,11 @@ files = [
 ]
 
 warn_unused_flag = '-Wno-unused-result'
+auxiliaries = []
 foreach file : files
-  executable(file, '@0@.c'.format(file),
+  auxiliaries += [executable(file, '@0@.c'.format(file),
     c_args: executable_cflags + (meson.get_compiler('c').has_argument(warn_unused_flag) ? [warn_unused_flag] : []),
     install: false,
     link_args: executable_linkflags
-  )
+  )]
 endforeach
