+++ test/unit/meson.build
@@ -1,12 +1,4 @@
 if get_option('enable_tests')
-  test_conf_data = configuration_data()
-  test_conf_data.set_quoted('RIZIN_BUILD_PATH', rizin_exe.full_path())
-  test_config_h = configure_file(
-    input: 'test_config.h.in',
-    output: 'test_config.h',
-    configuration: test_conf_data,
-  )
-
   subdir('auxiliary')
 
   tests = [
@@ -23,10 +15,8 @@ if get_option('enable_tests')
     'analysis_global_var',
     'analysis_xrefs',
     'annotated_code',
-    'autocmplt',
     'base64',
     'big',
-    'bin',
     'bin_lines',
     'bitmap',
     'bitvector',
@@ -44,9 +34,6 @@ if get_option('enable_tests')
     'debug',
     'debug_session',
     'diff',
-    'dwarf',
-    'dwarf_info',
-    'dwarf_integration',
     'ebcdic',
     'endian',
     'event',
@@ -72,15 +59,12 @@ if get_option('enable_tests')
     'json',
     'list',
     'ovf',
-    'pdb',
     'pj',
-    'project_migrate',
     'queue',
     'rbtree',
     'reg',
     'run',
     'rz_test',
-    'rzpipe',
     'serialize_analysis',
     'serialize_config',
     'serialize_debug',
@@ -116,9 +100,8 @@ if get_option('enable_tests')
 
   foreach test : tests
     exe = executable('test_@0@'.format(test), 'test_@0@.c'.format(test),
-      # '.' to find autogenerated "test_config.h"
       c_args: executable_cflags,
-      include_directories: [platform_inc, '.'],
+      include_directories: [platform_inc],
       dependencies: [
         rz_util_dep,
         rz_main_dep,
@@ -150,6 +133,6 @@ if get_option('enable_tests')
       implicit_include_directories: false,
       link_args: executable_linkflags
     )
-    test(test, exe, workdir: join_paths(meson.current_source_dir(), '..'), env: unit_test_env)
+    test(test, exe, workdir: join_paths(meson.current_source_dir(), '..'), env: unit_test_env, depends: auxiliaries, suite: 'unit')
   endforeach
 endif
